Better code organization
